# LoginSystemApi/DTOs/RegisterRequest.cs
namespace LoginSystemApi.DTOs
{
    public class RegisterRequest
    {
        public string Name { get; set; }
        public string Email { get; set; }
        public string Password { get; set; }
        public string CPF { get; set; }
        public string Role { get; set; }
    }
}